move ui related files from {core,driver} to ui
