+++ src/core/main.c
@@ -6,34 +6,36 @@
 #include <fcntl.h>
 #include <string.h>
 #include <pthread.h>
-#include "lvgl/lvgl.h"
-#include "self_test.h"
-#include "main_menu.h"
-#include "statusbar.h"
-#include "common.hh"
-#include "input_device.h"
-#include "osd.h"
-#include "thread.h"
-#include "imagesetting.h"
-#include "ht.h"
-#include "elrs.h"
-#include "../driver/hardware.h"
-#include "../driver/porting.h"
-#include "../driver/fans.h"
+
+#include <lvgl/lvgl.h>
+#include <minIni.h>
+
+#include "../bmi270/accel_gyro.h"
 #include "../driver/dm5680.h"
-#include "../driver/oled.h"
-#include "../driver/TP2825.h"
+#include "../driver/esp32.h"
+#include "../driver/fans.h"
+#include "../driver/hardware.h"
+#include "../driver/i2c.h"
 #include "../driver/it66021.h"
 #include "../driver/it66121.h"
 #include "../driver/mcp3021.h"
-#include "../driver/i2c.h"
-#include "../driver/esp32.h"
-#include "ui/page_scannow.h"
-#include "ui/page_power.h"
-#include "ui/page_source.h"
-#include "minIni.h"
-#include "../bmi270/accel_gyro.h"
+#include "../driver/oled.h"
+#include "ui/porting.h"
+#include "../driver/TP2825.h"
+#include "common.hh"
+#include "elrs.h"
+#include "ht.h"
+#include "input_device.h"
+#include "osd.h"
+#include "self_test.h"
 #include "thread.h"
+#include "thread.h"
+#include "ui/imagesetting.h"
+#include "ui/main_menu.h"
+#include "ui/page_power.h"
+#include "ui/page_scannow.h"
+#include "ui/page_source.h"
+#include "ui/statusbar.h"
 
 static void load_ini_setting(void)
 {
