+++ src/ui/imagesetting.h
