+++ src/ui/page_headtracker.c
@@ -1,7 +1,7 @@
 #include "page_headtracker.h"
 #include "page_common.h"
 #include "minIni.h"
-#include "style.h"
+#include "ui/style.h"
 #include "ht.h"
