+++ src/ui/page_record.c
@@ -1,7 +1,7 @@
 #include <stdlib.h>
 #include "page_record.h"
 #include "page_common.h"
-#include "style.h"
+#include "ui/style.h"
 #include "minIni.h"
 #include "playback.h"
 #include "../core/common.hh"
