+++ src/ui/page_scannow.c
@@ -13,17 +13,17 @@
 #include "fbtools.h"
 #include "msp_displayport.h"
 #include "page_common.h"
-#include "style.h"
+#include "ui/style.h"
 #include "../core/common.hh"
 #include "../core/defines.h"
-#include "../core/main_menu.h"
+#include "ui/main_menu.h"
 #include "../core/osd.h"
 #include "../driver/dm5680.h"
 #include "../driver/dm6302.h"
 #include "../driver/hardware.h"
 #include "../driver/i2c.h"
 #include "../driver/oled.h"
-#include "../driver/porting.h"
+#include "ui/porting.h"
 #include "../driver/uart.h"
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
