+++ src/ui/playback.c
@@ -11,7 +11,7 @@
 
 #include <log/log.h>
 
-#include "style.h"
+#include "ui/style.h"
 #include "ui/page_common.h"
 #include "mplayer.h"
 #include "common.hh"
