+++ src/ui/style.c
@@ -1,4 +1,4 @@
-#include "style.h"
+#include "ui/style.h"
 
 const lv_style_prop_t props[] = {LV_STYLE_BG_COLOR, 0};
 lv_style_transition_dsc_t transition_dsc;
